+++ src/main/java/com/twilio/rest/bulkexports/v1/export/DayReader.java
@@ -5,7 +5,7 @@
  *       /       /
  */
 
-package com.twilio.rest.supersim.v1;
+package com.twilio.rest.bulkexports.v1.export;
 
 import com.twilio.base.Page;
 import com.twilio.base.Reader;
@@ -20,49 +20,42 @@ import com.twilio.http.TwilioRestClient;
 import com.twilio.rest.Domains;
 
 /**
- * PLEASE NOTE that this class contains preview products that are subject to
- * change. Use them with caution. If you currently do not have developer preview
- * access, please contact help@twilio.com.
+ * PLEASE NOTE that this class contains beta products that are subject to
+ * change. Use them with caution.
  */
-public class SimReader extends Reader<Sim> {
-    private Sim.Status status;
-    private String fleet;
-    private String iccid;
+public class DayReader extends Reader<Day> {
+    private final String pathResourceType;
+    private String nextToken;
+    private String previousToken;
 
     /**
-     * The status of the Sim resources to read. Can be `new`, `active`, `inactive`,
-     * or `scheduled`..
+     * Construct a new DayReader.
      *
-     * @param status The status of the Sim resources to read
+     * @param pathResourceType The type of communication – Messages, Calls
+     */
+    public DayReader(final String pathResourceType) {
+        this.pathResourceType = pathResourceType;
+    }
+
+    /**
+     * The next_token.
+     *
+     * @param nextToken The next_token
      * @return this
      */
-    public SimReader setStatus(final Sim.Status status) {
-        this.status = status;
+    public DayReader setNextToken(final String nextToken) {
+        this.nextToken = nextToken;
         return this;
     }
 
     /**
-     * The SID or unique name of the Fleet to which a list of Sims are assigned..
+     * The previous_token.
      *
-     * @param fleet The SID or unique name of the Fleet to which a list of Sims are
-     *              assigned
+     * @param previousToken The previous_token
      * @return this
      */
-    public SimReader setFleet(final String fleet) {
-        this.fleet = fleet;
-        return this;
-    }
-
-    /**
-     * The [ICCID](https://en.wikipedia.org/wiki/Subscriber_identity_module#ICCID)
-     * associated with a Super SIM to filter the list by. Passing this parameter
-     * will always return a list containing zero or one SIMs..
-     *
-     * @param iccid The ICCID associated with a Super SIM to filter the list by
-     * @return this
-     */
-    public SimReader setIccid(final String iccid) {
-        this.iccid = iccid;
+    public DayReader setPreviousToken(final String previousToken) {
+        this.previousToken = previousToken;
         return this;
     }
 
@@ -70,10 +63,10 @@ public class SimReader extends Reader<Sim> {
      * Make the request to the Twilio API to perform the read.
      *
      * @param client TwilioRestClient with which to make the request
-     * @return Sim ResourceSet
+     * @return Day ResourceSet
      */
     @Override
-    public ResourceSet<Sim> read(final TwilioRestClient client) {
+    public ResourceSet<Day> read(final TwilioRestClient client) {
         return new ResourceSet<>(this, client, firstPage(client));
     }
 
@@ -81,15 +74,15 @@ public class SimReader extends Reader<Sim> {
      * Make the request to the Twilio API to perform the read.
      *
      * @param client TwilioRestClient with which to make the request
-     * @return Sim ResourceSet
+     * @return Day ResourceSet
      */
     @Override
     @SuppressWarnings("checkstyle:linelength")
-    public Page<Sim> firstPage(final TwilioRestClient client) {
+    public Page<Day> firstPage(final TwilioRestClient client) {
         Request request = new Request(
             HttpMethod.GET,
-            Domains.SUPERSIM.toString(),
-            "/v1/Sims",
+            Domains.BULKEXPORTS.toString(),
+            "/v1/Exports/" + this.pathResourceType + "/Days",
             client.getRegion()
         );
 
@@ -102,11 +95,11 @@ public class SimReader extends Reader<Sim> {
      *
      * @param targetUrl API-generated URL for the requested results page
      * @param client TwilioRestClient with which to make the request
-     * @return Sim ResourceSet
+     * @return Day ResourceSet
      */
     @Override
     @SuppressWarnings("checkstyle:linelength")
-    public Page<Sim> getPage(final String targetUrl, final TwilioRestClient client) {
+    public Page<Day> getPage(final String targetUrl, final TwilioRestClient client) {
         Request request = new Request(
             HttpMethod.GET,
             targetUrl
@@ -123,12 +116,12 @@ public class SimReader extends Reader<Sim> {
      * @return Next Page
      */
     @Override
-    public Page<Sim> nextPage(final Page<Sim> page,
+    public Page<Day> nextPage(final Page<Day> page,
                               final TwilioRestClient client) {
         Request request = new Request(
             HttpMethod.GET,
             page.getNextPageUrl(
-                Domains.SUPERSIM.toString(),
+                Domains.BULKEXPORTS.toString(),
                 client.getRegion()
             )
         );
@@ -143,12 +136,12 @@ public class SimReader extends Reader<Sim> {
      * @return Previous Page
      */
     @Override
-    public Page<Sim> previousPage(final Page<Sim> page,
+    public Page<Day> previousPage(final Page<Day> page,
                                   final TwilioRestClient client) {
         Request request = new Request(
             HttpMethod.GET,
             page.getPreviousPageUrl(
-                Domains.SUPERSIM.toString(),
+                Domains.BULKEXPORTS.toString(),
                 client.getRegion()
             )
         );
@@ -156,17 +149,17 @@ public class SimReader extends Reader<Sim> {
     }
 
     /**
-     * Generate a Page of Sim Resources for a given request.
+     * Generate a Page of Day Resources for a given request.
      *
      * @param client TwilioRestClient with which to make the request
      * @param request Request to generate a page for
      * @return Page for the Request
      */
-    private Page<Sim> pageForRequest(final TwilioRestClient client, final Request request) {
+    private Page<Day> pageForRequest(final TwilioRestClient client, final Request request) {
         Response response = client.request(request);
 
         if (response == null) {
-            throw new ApiConnectionException("Sim read failed: Unable to connect to server");
+            throw new ApiConnectionException("Day read failed: Unable to connect to server");
         } else if (!TwilioRestClient.SUCCESS.apply(response.getStatusCode())) {
             RestException restException = RestException.fromJson(response.getStream(), client.getObjectMapper());
             if (restException == null) {
@@ -183,9 +176,9 @@ public class SimReader extends Reader<Sim> {
         }
 
         return Page.fromJson(
-            "sims",
+            "days",
             response.getContent(),
-            Sim.class,
+            Day.class,
             client.getObjectMapper()
         );
     }
@@ -196,16 +189,12 @@ public class SimReader extends Reader<Sim> {
      * @param request Request to add query string arguments to
      */
     private void addQueryParams(final Request request) {
-        if (status != null) {
-            request.addQueryParam("Status", status.toString());
+        if (nextToken != null) {
+            request.addQueryParam("NextToken", nextToken);
         }
 
-        if (fleet != null) {
-            request.addQueryParam("Fleet", fleet.toString());
-        }
-
-        if (iccid != null) {
-            request.addQueryParam("Iccid", iccid);
+        if (previousToken != null) {
+            request.addQueryParam("PreviousToken", previousToken);
         }
 
         if (getPageSize() != null) {